Update styles.css
after much consideration and trying out different approaches, ultimately the mirrored text just did not fit the look I was going for. And so instead I decided to opt for curved text on my next session.

for now, we have deleted the mirrored text and cornered letter styles. And also have swapped out the colors currently being used on our corner circles to a purple, teal, and green so that it is not so harsh on the eyes.